# Sistema_Gestión_Donante/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Gestión de Donantes - API",
        "description": "API para el módulo de donantes del sistema de adopciones de animales",
        "contact": {
            "email": "admin@sistema-adopciones.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/api/mascotas-donadas": {
            "get": {
                "tags": [
                    "Mascotas Donadas"
                ],
                "summary": "Obtener todas las mascotas donadas",
                "operationId": "48abc94121acb5491a517edf7ef1f488",
                "responses": {
                    "200": {
                        "description": "Lista de mascotas donadas obtenida exitosamente"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Mascotas Donadas"
                ],
                "summary": "Crear una nueva mascota donada",
                "operationId": "fb77616f93452eff162dc1bccac8d012",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "donante_id",
                                    "mascota_id",
                                    "fecha_donacion",
                                    "motivo_donacion"
                                ],
                                "properties": {
                                    "donante_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "mascota_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "fecha_donacion": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-06-08"
                                    },
                                    "motivo_donacion": {
                                        "type": "string",
                                        "example": "No puedo cuidarla más"
                                    },
                                    "estado_revision": {
                                        "type": "string",
                                        "example": "Pendiente"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mascota donada creada exitosamente"
                    },
                    "400": {
                        "description": "Datos inválidos"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            }
        },
        "/api/mascotas-donadas/{id}": {
            "get": {
                "tags": [
                    "Mascotas Donadas"
                ],
                "summary": "Obtener una mascota donada por ID",
                "operationId": "ddedca5f3f07487521f1d5fef5a100c6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la mascota donada",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mascota donada obtenida exitosamente"
                    },
                    "404": {
                        "description": "Mascota donada no encontrada"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            },
            "put": {
                "tags": [
                    "Mascotas Donadas"
                ],
                "summary": "Actualizar una mascota donada existente",
                "operationId": "181862286087593044bfc3fb1286cc18",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la mascota donada",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "donante_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "mascota_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "fecha_donacion": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-06-08"
                                    },
                                    "motivo_donacion": {
                                        "type": "string",
                                        "example": "Motivo actualizado"
                                    },
                                    "estado_revision": {
                                        "type": "string",
                                        "example": "Aceptada"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mascota donada actualizada exitosamente"
                    },
                    "400": {
                        "description": "Datos inválidos"
                    },
                    "404": {
                        "description": "Mascota donada no encontrada"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Mascotas Donadas"
                ],
                "summary": "Eliminar una mascota donada",
                "operationId": "1121265361316ae95d28ec6fff6cad49",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la mascota donada",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mascota donada eliminada exitosamente"
                    },
                    "404": {
                        "description": "Mascota donada no encontrada"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            }
        },
        "/api/mascotas-donadas/donante/{donanteId}": {
            "get": {
                "tags": [
                    "Mascotas Donadas"
                ],
                "summary": "Obtener mascotas donadas por donante",
                "operationId": "0853172f0d64012c092d7b079dde6e52",
                "parameters": [
                    {
                        "name": "donanteId",
                        "in": "path",
                        "description": "ID del donante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de mascotas donadas por donante obtenida exitosamente"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            }
        },
        "/api/mascotas-donadas/estado/{estado}": {
            "get": {
                "tags": [
                    "Mascotas Donadas"
                ],
                "summary": "Obtener mascotas donadas por estado",
                "operationId": "2ecdcfecd0d19bd40042731c0aa32df7",
                "parameters": [
                    {
                        "name": "estado",
                        "in": "path",
                        "description": "Estado de revisión",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de mascotas donadas por estado obtenida exitosamente"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            }
        },
...
    },
...
}